Define (morally) fmap for Progress and use it
The use is questionable, will be eliminated if we keep using the simple @rpc@.
But it'll quite likely be useful elsewhere. BTW, I think Progress is better
with its two arguments (even if the most useful Functor instance can't be
defined for it), because it can be used in a type-safer way by the users
of our library.