Turn SomeException to a string early, to avoid serialising it
I have no idea how to serialise SomeException. I can't even throw it
and catch it at the right place, because it crosses process boundaries.
Unless we really need it, let's try to analise and pretty-print
exceptions in the runGHC thread already. I can take a handler that does that,
if we need generality.