Record original exception in ExternalException
This means we can show parse errors if the server returns something weird.  The
tricky part is that it's hard to distinguish between a parse error because the
server crashed and a parse error because the server did something bold (like
output to stdout). We can't simply check if the server process is still running
because that would make the errors reported non-deterministic (depending on
whether the OS cleaned up the process yet or not).