Move error collectio into the Ghc monad; exception handling stays outside
Moreover, exception handling now has no access to the already
collected errors and this results in worse error reporting.
This can be fixed by creating the errors IORef in the IO monad
where we catch exceptions and then passing it as an argument
to the Ghc monad code. The IORef would need to be emptied
at each entrace to the Ghc monad. But perhaps there's a better way.