Fix the use of handleSourceError
Some compilation errors are reported as free-form exceptions (e.g., CPP errors),
yet other compilation errors are presented as SourceError, but escape
the log_action handler. However, unlike the free-form exceptions,
they can be captured inside the Ghc monad, which makes it possible
to handle them and keep partial compilation results intact in the Ghc monad
state. Function handleSourceError captures such SourceError exceptions.
If we tried really hard we could call collectSrcError on the SourceError values,
but for now we just package them in the OtherError constructor
(instead of just printing them, as was the case before this commit).