Fix weird numbering problem
For some reason, it seems that compileInGhc occassionally uses a handler from a
previous invocation. I don't know why this happens, but we can work around it
by making sure that we always use the *same* handler. For the particular
problem we were having, that means creating an IORef which holds the number of
the file once and for all, rather than creating a new IORef on every request,
and just resetting that IORef to 1 at the start of the request.