Attempt to work around the timestamp+recompilation GHC API problem
The problem is rounding of file timestamps to 1 second for the purpose
of deciding if a file needs to be recompiled. With fast enough user file
changes or any automatic changes (say, refactoring), this breaks recompilation.
Unfortunately CPP preprocessing does not seem to work with StringBuffers,
so the workaround in this commit is probably useless.