Work around low resolution of getModificationTime
ghc uses getModificationTime in order to assess whether files have changed;
since getModificationTime has a second-resolution, this does not always work in
the context of IdeSession, where multiple changes may be made to a module
within the same second.

We now explicitly set the modification time of the files we create manually,
using a "logical timestamp", which we simply increment on every call to
ModilePut or ModuleSource. This guarantees that when we overwrite files (and
only when we overwrite files) their timestamp is incremented.

Note that if the user explicitly sets ModulePut for a module which hasn't
changed it will be recompiled. If we don't want that, we need to implement some
sort of hashing.