Make sure all sessions are properly closed
To do this, I've split 'testSingle' into a number of separate funtions:

  withSession :: SessionConfig -> (IdeSession -> IO a) -> IO a

    The usual bracketing around creating and shutting down a session

  withConfiguredSession :: [String] -> (IdeSession -> IO a) -> IO a

    Like withSession, but configuring the session to use a temp dir

  loadModulesFrom :: IdeSession -> FilePath -> IO ()

    To load all modules from the given directory into the session

testSingle is gone completely: it's now two calls: withConfiguredSession and
loadModulesFrom. This makes it clearer what's going on in the tests, and
moreover means that for tests where we don't want that inital call to update we
can just leave it out.

The tests now run without broken pipes errors.