Change test to catch duplicated module names
Previously we experienced GHC complains due to a module name not matching
file name, but I can't reproduce it now, so I'm removing the old test
and introducing a related one. These situations are not very relevant
to our code, anyway, though it's good to know that we handle the resulting
exceptions in a proper way, etc.