Insert a check that no unneeded recompilation happens during the test
or, rather, that the progress counter does not grow too high.
To check that it's not too low, we'd need to provide another number,
indicating the minimal number of modules that need be recompiled.
It wouldn't be strict at all, because we never know if an erronous
module is the first or the last to be attempted.