Generalize RPC model
We started with one request-one response, then we moved to one request-multiple
responses, and now we generalize to general conversations (multiple requests
and responses within one conversations). It is still useful to mark
conversations as such (rather than just provide a non-scoped get/put methods)
because it marks points where the client communicates with the server; if
exceptions happen during a conversation, the server is in an ill-defined state
and is therefore terminated; but exceptions outside of the conversation do not
affect the server.