Fix 'blocked on MVar'
rpcServer spawns three threads, one to read requests, one to write responses,
and one to run the server thread itself (the bit that you write, basically).
the bug happened if one of those threads before the other got a chance to
install an exception handler. I knew that this might happen, but I thought this
was a theoretical possibility only. Apparently not!