Clarify types
The innocent summary "clarify types" actually hides a rather large refactoring.
We take advantage of the increased flexibility (in types) of the RPC server to
clarify the types of the RPC calls. This in turn made it possible to
simplify/clarify a lot of other code, as well as the tests.