Fix subtle race condition in ghcHandleRun
If the thread that throws the UserInterrupt threw this interrupt before runStmt
then it would be caught by the surrounding ghandleJust, but as a result that
thread would not be terminated. There is however a very simple fix: we can just
exit that thread when we receive an interrupt (no point interrupting more than
once).