Fix updating session with data files and fix a test with .h file
The test checks that updating session with GHC pragmas works,
but the test also uses a .h file, which in the new semantics od module names
is ignored as a source file. Here we use a hacky workaround, adding the .h
file as a data file, testing data-file addition in the process.