Introduce IdeUpdateResult. Start using lenses
We want to return more info from the session updates. So rather than returning
thew new IseIdleState, the session updates now return an IdeUpdateResult,
currently defined simply as

    data IdeUpdateResult = IdeUpdateResult {
         _updatedState :: IdeIdleState
       }

Clearly, this doesn't add any new fields just yet, but we will soon. However,
this introduces yet another layer of records. With the introduction of
`ManagedFields` we were already nesting records two-deep, and now three-deep.
This means that the {..} (and related) syntax we were using to access and
(especially) update records doesn't work so well anymore. For that reason, we
now use lenses, and I have modified the type of `IdeSessionUpdate` to use the
state monad:

    newtype IdeSessionUpdate = IdeSessionUpdate {
        runSessionUpdate :: SessionConfig -> StateT IdeUpdateResult IO ()
      }

This allows us to write

    updateModuleDelete :: ModuleName -> IdeSessionUpdate
    updateModuleDelete m = IdeSessionUpdate $ \ideConfig -> do
      liftIO $ removeFile (internalFile ideConfig m)
      modify (updatedState .> ideManagedFiles .> sourceFiles) (delete m)

instead of

    updateModuleDelete :: ModuleName -> IdeSessionUpdate
    updateModuleDelete m =
        IdeSessionUpdate $ \ideConfig state@IdeIdleState{ideManagedFiles} -> do
          removeFile (internalFile ideConfig m)
          let sF = delete m (sourceFiles ideManagedFiles)
          return state {ideManagedFiles = ideManagedFiles {sourceFiles = sF}}

Note that in the "before" example it is already awkward to update the
'sourceFiles' field (nested two deep), it would be even more awkward with the
new `IdeUpdateResult`. With lenses however this is much simplified.