Fix race condition/better treatment of GhcRunInterrupt
It might be that the GhcRunDone might cross a GhcRunInterrupt request. The
client now responds GhcRunDone with GhcRunAckDone and the server waits to
receive this GhcRunAckDone; moreover, we guarantee that GhcRunAckDone is the
last message we sent for any particular runStmt. This avoids the race
condition.

Also change the handling of GhcRunInterrupt in a more principled way. This was
necessary as part of the fix to the race condition, and also seems to have
fixed the weird exception problem we were having in isolation-runner.