Fix 'blocked indefinitely' in resetSession tests
This was caused by a 'runWait' on a RunActions from the obsolete server, which
was waiting indefinitely for a response. This now throws an exception ("runWait
force-cancelled") -- one call at least; in order to implement this, introduced
'forceCancel' on 'RunActions'. This is not exported, however, which broke the
nice standalone-i-ness of RunAPI, so merged that back with the rest
unfortuately.

Also start using Async; still need to do that in various other places too.