(Partially) deal with hard ghc crash in runWait
We should test hard crashes more thoroughly though, and decide on a policy of
whether or not we should restart the session on a RunGhcException.

Tracked the ghc exception itself down to a bug in the GHC API:
http://hackage.haskell.org/trac/ghc/ticket/7539

Also modified the code so that we can easily get line numbers in "blocked
indefinitely" exceptions (just set DEBUGGING to 1 in BlockingOps).