Introduce ide-backend RTS. Closes #24.
The problem with #24 was already described in the issue; we now load a module
'IdeBackendRTS' into the ghc session on start up. As part of the RTS we define
a function 'run' which, amongst other things, resets the stdout, stderr and
stdin handles. This rather low-level coding though, so may be a bit fragile to
changes in the ghc RTS. Until we have proper GHC support for this, however, I
don't think there is a way around that.

I'm not merging this with master just yet because there is one integration test
failing ("Maintain list of compiled modules"). I looked at this for a while but
it seems a genuine bug somewhere other than the new RTS. I can't make sense of
it yet.