Add a CPP flag to Cabal and ifdef the GHCI defendencies workaround precisely
Citing from https://github.com/fpco/ide-backend/wiki/Ghc-bugs:

* GHC API dependency analysis is broken

Worked around for 7.4.2 (with some luck it should work for any 7.4.*), workaround ifdefed to be ignored for GHC >= 7.6.2 and to cause compilation failure for GHC == 7.6.1. It is fixed upstream for 7.6.2 as GHCi erroneously unloads modules after a failed :reload. If we ever switch to 7.6.1, a similar workaround has to be cloned from GHC 7.6.1 sources and patched with the 7.6.2 fix.