Add a test reproducing the import+unload problem reported by snoyberg
I appears GHC does not unload modules if the error is in the import section.
Possibly, the reasoning goes that if module dependencies are wrong then
the module name can be wrong as well, so it's not clear which module to unload
(GHC does not equate file names with module names and not always checks
that they match). Will investigate further.