Alternative approach
It seems some information is lost after typechecking -- in particular, the
global reader environment that we were consulting appears not to be useful
anymore (the comment in ghc says that this environment is used "during
renaming", confirming this suspicion). So we now approach the problem slightly
different. We first walk the AST after renaming, collecting all info except
types; then we walk it again after type checking, but now just amending the
information already collected in the first phase and only adding type
information. That is, we only consult the GblRdrEnv when we walk the renamed
(but not yet type checked) AST.

However, when we run typecheck-dir on the Cabal source, there are warnings
about missing GblRdrEnv entries. I don't yet know why.