Fix the 'missing entries' reported in a92c933
The GlbobalRdrEnv returned by the type checker contains may contain multiple
entries for OccNames. So, if we do

  import qualified Data.ByteString.Lazy as BS

then the environment contains *two* entries for the occname "map". The comment
in the ghc source says that there are only multiple entries in case of
collisions, but doesn't seem to be the case, since there is no collision in the
above example. Either way, to look up the right entry we now

  given a Name n,
    lookup the list of entries using the associated OccName
    filter that list to find an entry that corresponds to n

This fixes the problem. We no longer have any entries with missing info when
typechecking all of Cabal, with the exception of assert(Error).