Fix pretty-printing of types
Before the type of

    foo (x, y) = x

was reported as

    (t, t) -> t

now it's (almost) correctly reported as

    (t, t1) -> t1

I'm saying "almost" because we might really want a forall here -- but on the
other hand, a forall limits the scope of the type variable, making it slightly
unclear what we should when reporting the type of 'x' in the body, say. The
"Type information" tests explore some of these cases (but currently pass with
the above).