Determine which modules have changed in this GHC call
The only way I know to do that is using the new GHC plugin mechanism.
Keeping timestamps of the previous call and comparing timestamps
of files would tell us which files have changed, but in general we don't
have a mapping from files to modules. From what I see GHC doesn't expose
it's own mechanism for deciding which files to recompile or which modules
have been changed recently.