Move the IdInfo cache to GhcHsWalk
This is WIP: the IdMaps do not yet contain keys into the cache (only the
autocomplete maps do) and a problem emerged caching IdInfos for IdMap
for different modules (see the result of the "Type information" tests).
Either we should cache per-module or don't cache occurences
of imported ids, just as we don't cache binding occurences now.