Major refactoring
The real feature addition is that we now share filepaths. However, completely
redesigned how we deal with explicit sharing, and are we now a lot more
principled about it. Most types now come in Foo and XFoo form, where XFoo is
the version that has explicit sharing, plus a type class

  normalize :: Normalize a b => ExplicitSharingCache -> a -> b

We now carry these 'denormalized' types all the way down to the ide session
level.