Major refactoring
Things were getting messy; this is the start if a major tidying up round.

The most important goal of this round (and we're not quite there yet) is that
we make a clear distinction between internal types (IdeSession.Types.Private)
that use explicit sharing and moreover make use of specialized types
(IdeSession.Strict.*) that guarantee the right strictness properties, and
public types (IdeSession.Types.Public). By splitting all this stuff up into
separate modules we can guarantee that we only import that types we intend:
private types for internal code, public types for the public API.