SessionConfig specifies whether to generate per-module type/autocompletion info
This is specified at session config and can't be changed, because we'd need
to recompile all modules when the option is turned on. This is mostly
equivalent to starting a new session, but would complicate code.

TODO: perhaps make it safer by wrapping the per-module info types in Maybies
TODO: perhaps make it more fine-grained, e.g., set LoadedModules separately