Fix test failure 'concurrentGetPut'
The test which failed very occasionally has a server which waits for one
request and replies once. the client sends one request, waits for the server to
respond, and exits. that all works fine, except that the server might exit
sometimes thereby taking down the pipe before the client gets a chance to read
from it.  changing the server to look forever (server an infinite number of
requests) fixes the issue; another solution would be for the client to tell the
server "it's ok to exit now" and then make sure not to wait for another
response from the server.