Refactor ghcHandleCompile
The immediate objective of the refactoring is to make sure getLoadedModules
works even when configGenerateModInfo is False; but in the process cleaned up
internal data structures, refactor ghcHandleCompile to be a lot more readable,
avoid recomputing data that we don't need to recompute, and improve memory
consumption of the server (which was hanging on to autocompletion data for no
discernable reason).

Not all unit tests pass; committing so that the code can be reviewed.