Refactor ghcHandleCommit again
The old refactoring did not take into account that we might have info for
modules that are not "loaded" as such because they have type errors; now it
becomes trickier to figure out which modules got "unloaded" vs "removed
completely" or anything in between. To solve this problem properly, we now
cache a little more information. The new version should also be faster, as it
does a strictly linear scan over the old cached information and the new module
summaries that we get from ghc, with no searching involved.

This closes #44.