Initial attempt at Issue #50
We record the subexpression types for identifiers, application, lambda
abstraction, (overloaded) literals, and let expressions. The approach seems to
work well and is efficient (every node in the AST is considered only once).
Just left to do is to add support for more expressions, and the order of the
"dominators" seems somewhat haphazard at the moment.