Deal with errors on ghc server init
The previous commit exposed a long latent bug: ghc initialization might fail
immediately, but we would only detect this on the first call to updateSession;
we now detect this immediately, and also deal with a potential infinite restart
cycle (we only attempt to restart once on updateSession). Note that this means
that getSourceErrors may return a non-empty list even before the first session
update (for instance, when some flags that are passed to ghc are invalid).