Implement a proper handshake
We were passing some options as command line arguments, others as separate RPC
calls. Now we do a single RPC exchange when the server starts, but before we
start the ghc thread. This makes it possible to pass stuff to the server which
is needed _before_ starting ghc (such as the PackageDBStack) without hacks;
previously, we needed to somehow encode these and pass them as command line
arguments.

This paves the way for #158.