Refactor how we set ghc settings
We want updateGhcSettings to be stateless. This means that it doesn't make much
sense to have some static options in SessionConfig and others in the
IdeSessionState. This refactoring paves the way for #115, but doesn't change
anything fundamentally yet. It just refactors what's already there and deals
with the fallout in the tests.