Set flags for compilation only when needed
This brings down apparent time (and especially memory) complexity
of cases where each module is loaded via a separate update operation from
O(number_of_modules ^ 2) to O(number_of_modules * number_of_flag_changes)
(not very scientifically measured and only for the range of up to 500 modules)
where the flags are changed, e.g., every time compilation and running snippets
are interleaved.

TODO: the increased memory usage when flags are changed suggests
a memory leak in GHC API when caches are invalidated due to flags change
or another problem. May be worth investigating if the scenario of, e.g.,
multiple interleaved recompilation and running turns out common enough.
Or we may try to avoid flags change in that case (the verbosity flag),
but it's rather hard, since the flag change is a work around another
reported GHC bug.

See also
https://ghc.haskell.org/trac/ghc/ticket/7478
(the fix to) which is the cause of the slowdown reported in #167
(the caches are invalidated, but there weren't before, but that was buggy).