Don't invoke progress callback in buildExe and buildDoc
It wasn't particularly informative, anyway (#120) and now we are going
to call the procedures via a separete exe, not to capture stdout/stderr
of the whole process (#180), so passing the callback is not possible.
A proper implementations would count the number of modules
submitted for building and then scan the stdout log file
and signal progress whenever a suitable line (or any line) is emitted.