Add runExe tests for each previous readProcess test
The results agree. Note that runExe and runStmt outputs differ
in case of an exception (e.g., in "Switch from one to" tests,
where 'error' is called), probably because runExe doesn't catch
exceptions, so their handlers can, e.g., print something to stderr.
This is reflected in the types to an extent: 'RunActions RunResult'
for runStmt vs 'RunActions ExitCode' for runExe.