Refactor updateSession, part 1
The logic for updateSession was getting convoluted due to the need for
restarting at arbitrary points, which was making further changes to
updateSession harder (in particular, modifying it so that we can do updates to
data files for #176). In this first step of the refactoring we change the
datatype representing updates, but don't yet take advantage of this in
updateSession itself.