Fix race condition with new RPC infrastructure
Opening a named pipe for writing without a corresponding reader will result in
a "Device not configured" exception. That's why we opened all read ends of
pipes first, and then opened the write ends in blocking mode -- making sure to
open the read ends before the write ends to avoid a deadlock.

However, it turns out that when you call hGetSome on a named pipe which has
been opened for reading but that has not yet been opened for writing, hGetSome
will return EOF. This explains the non-determinism we were seeing: there was a
race condiiton between the server opening the pipe for writing, and the client
doing the first call to hGetSome (and concluding that the connection was
severed because it got EOF).

We now use a blocking opening for _both_ reading and writing (and now opening
the pipes in the same order on both sides, to avoid deadlock again). This seems
to resolve the issue.