Deal with force termination of snippets
When executing snippets under forkProcess rather than forkIO, some snippets
that were previously interruptible with an asynchronous exception are no
longer; a simple black hole is one example

    loop = loop

but this is an example too:

    loop = forever $ print 1

This first one I can _kind of_ see why it might happen; the second one I have
no idea. Either way, we need to be able to force terminate snippets, and
moreover make sure that when we terminate the entire session any running
snippets get terminated too (potentially forcefully if we are force-terminating
the session).