Delay async exceptions in runStmt (#219)
Because of potential deadlocks in the spawned snippet server, calls to runStmt
after often wrapped in a call to timeout. However, if this timeout expires
during communication with the main server, the whole session becomes unuseable.

To avoid this, we now mask async exceptions in the communication with the main
server. Of course, this has some dangers of its own. Should discuss.