Add tests for #218
The 2 tests are added, but not included in test runs, until the issue
is fixed.

Should this actually go into TestSuite.Tests.Issues? I think it is a rather
basic FFI functionality, not a tricky corner case or a trigger for a GHC bug,
so perhaps TestSuite.Tests.FFI makes sense?