Avoid unnecessary double restart on server crash
When the server died we used to restart it with the _current_ settings, and
then re-apply the session update; but if this session update itself requires a
session restart, we would end up requiring a _second_ restart. We now make sure
to restart the server right away with the right parameters.