Hide the existence of ide-backend-rts
This is an implementation of the idea discussed in #289:
to embed in ide-backend-server's binary a properly compiled and
packaged version of its rts library, so that clients need not worry
anymore about having that special dependency installed somewhere the
server can find it.

The important changes are in ide-backend-server's Setup.hs script,
that now must also build and pack the rts in advance (moved to
ide-backend-server's source dir), ensuring:

  1. that exactly the same version of ghc is used, and

  2. that it is packed as a relocatable package (so that it
     doesn't matter where the server puts it at runtime).

We ensure 1) as follows: whenever "Setup.hs configure" is run, we
also run "Setup.hs configure" inside the rts directory, passing the
same command line parameters we received (overriding the installation
directory). For 2), we cannot rely on Cabal (not yet, at least,  support
is coming in Cabal-1.24, it looks), so we first use Cabal to package the
rts using absolute paths and then manually insert the required "${pkgroot}"
in the "installed package configuration".