Receive location of ide-backend-{server/exe-cabal}
Allow clients to indicate how to find the programs
ide-backend-server and ide-backend-exe-cabal via new fields
in `SessionConfig`. The `defaultSessionConfig` mimics the old
behaviour: to look for a program named ide-backend-server
(resp. -exe-cabal) on the system path (but any client that previously
relied on setting an appropriate path in 'configExtraPathDirs' will
have to be updated).